requirements.txt
fastapi
uvicorn
tensorflow-cpu
scikit-learn
pillow